+++ packages/transaction-pay-controller/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add optional `requiredTransactionCount` to Relay-backed `TransactionPayQuote` values so clients can predict how many TransactionController child transactions a quote will create ([#9897](https://github.com/MetaMask/core/pull/9897))
+
 ### Changed
 
 - Bump `@metamask/assets-controller` from `^13.1.2` to `^13.1.4` ([#9873](https://github.com/MetaMask/core/pull/9873), [#9886](https://github.com/MetaMask/core/pull/9886))